feat(projects): Reworked requests for IAM

+++ src/components/Project/Workspace.tsx
@@ -6,7 +6,7 @@ import { getUsers } from '../../api/gatewayClientAPI'
 import {
 	addUserToProject,
 	deleteProject,
-	getProjects,
+	getProjectsForUser,
 	removeUserFromProject,
 } from '../../api/projects'
 import { User } from '../../api/types'
@@ -24,16 +24,12 @@ const ProjectDashboard = () => {
 	const { trackEvent } = useMatomo()
 
 	const modalRef = useRef<ModalComponentHandle>(null)
-	const [users, setUsers] = React.useState<User[]>([])
 	const {
 		user: [user],
-		projects: [_, setProjects], // eslint-disable-line @typescript-eslint/no-unused-vars
+		users: [users],
+		userProjects: [_, setProjects], // eslint-disable-line
 		selectedProject: [project, setProject],
 	} = useAppStore()
-
-	useEffect(() => {
-		getUsers().then(users => setUsers(users))
-	}, [])
 
 	const handleAddUserToProject = (userId: string) => {
 		if (!project?.name) return
@@ -89,12 +85,12 @@ const ProjectDashboard = () => {
 			'Permanently remove this project ?'
 		)
 
-		if (reply) {
+		if (reply && user?.uid) {
 			deleteProject(projectName)
 				.then(res => {
 					showNotif('Project deleted', 'success')
 
-					getProjects().then(projects => {
+					getProjectsForUser(user?.uid ?? '').then(projects => {
 						setProjects(projects)
 					})
 					navigate(`${ROUTE_PREFIX}/projects`)
@@ -139,7 +135,7 @@ const ProjectDashboard = () => {
 						}}
 					>
 						<Box sx={{ gridColumn: '1', gridRow: '1' }}>
-							{project && (
+							{project && users && (
 								<ProjectCard
 									project={project}
 									handleRemoveProject={handleRemoveProject}
@@ -150,12 +146,14 @@ const ProjectDashboard = () => {
 
 						<>
 							<Box sx={{ gridColumn: '2', gridRow: '1' }}>
-								<MemberCard
-									project={project}
-									users={users}
-									handleAddUserToProject={handleAddUserToProject}
-									handleRemoveUserFromProject={handleRemoveUserFromProject}
-								/>
+								{users && (
+									<MemberCard
+										project={project}
+										users={users}
+										handleAddUserToProject={handleAddUserToProject}
+										handleRemoveUserFromProject={handleRemoveUserFromProject}
+									/>
+								)}
 							</Box>
 							<Box sx={{ gridColumn: '1', gridRow: '2' }}>
 								{/* <Tools /> */}
